requirements.txt
attrs==23.1.0
brotlipy==0.7.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama==0.4.6
coloredlogs==15.0.1
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
ddddocr==1.4.7
exceptiongroup==1.1.2
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
flatbuffers==23.5.26
h11==0.14.0
humanfriendly==10.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata==6.7.0
mpmath==1.3.0
numpy==1.21.6
onnxruntime==1.14.1
opencv-python-headless==4.8.0.74
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
protobuf==4.23.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
requests @ file:///C:/ci/requests_1657735288441/work
requests-file @ file:///Users/ktietz/demo/mc3/conda-bld/requests-file_1629455781986/work
selenium==4.11.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.10.1
tldextract @ file:///opt/conda/conda-bld/tldextract_1646638314385/work
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wsproto==1.2.0
zipp==3.15.0
